fix: GoFeatureFlagUser does not serialize
System.Text.Json does not serialize private members of a class.
Since the class GoFeatureFlagUser had the following private
members:
```
        private string Key { get; set; }
        private bool Anonymous { get; set; }
        private Dictionary<string, object> Custom { get; set; }
```
When it was serialized as the User in `GOFeatureFlagRequest`, the Json
that was returned was `{ "user": {},  ...` instead of
`{ "user": { "key: :  ...`

This commit fixes this issue by making the property public and
adding a test.

Signed-off-by: Eric Holton <eric.holton@gmail.com>